pyproject.toml
[project]
name = "mail-spoof-audit"
version = "0.1.0"
description = "CLI tool to audit SPF, DMARC, DKIM and spoofing risks across multiple TLDs"
authors = [{name = "mvivier001"}]
dependencies = ["dnspython"]
requires-python = ">=3.8"

[project.scripts]
mail-spoof-audit = "mail_spoof_audit.cli:main"
